Drop various debugging bits, add clarifying comments
